requirements.txt
fastapi==0.115.0
psycopg2==2.9.10
SQLAlchemy==2.0.35
uvicorn==0.30.6
pydantic[email]==2.9.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
pydantic-settings==2.5.2
bcrypt==4.2.0
